fix: notify dApp if Tezos operation failed in Wallet
Added gentle reject if there are no wallets.
The function approveTezosRequest was updated to
gently reject the request if there are no Tezos
wallets available.
Previously the problem of not initialized wallets
was leading to an exception on Wallet size and
an endless waiting on dApp side.

The Wallet also waits for exceptions from Tezos send operation.
They are reported to dApp side.