feat: Updated dApp modal to show operation description
Most Tezos operations use "tezos_send" as the operation type.
Inside it, the "kind" field is used to specify the operation kind, like
"transaction", "origination", etc. So distinguising the operation
contents is critical for Tezos.

The dApp now shows the operation data as a description (which is
optional) under the list of operations in the dApp.
The $(adderss) placeholder will be filled with the address of the
current account for Tezos operations.